==> 03_Notebooks/02_Development/01_setup.ipynb <==
# SET UP

This notebook sets up the project environment and directory structure for the **SmartPort Delay Risk Scoring** project. 

It defines the project environment, paths, imports required libraries, and loads the initial dataset of container ship movements.

It also creates the training, validation, and working samples that will be used in subsequent notebooks for data quality checks, feature engineering, modeling and production code preparation. The goal is to establish a clear and organized foundation before beginning data quality, transformation and modeling steps.

## PROJECT ENVIROMENT

### Create and activate environment
conda create -n smartport python=3.10
conda activate smartport

### Install dependencies
pip install -r requirements.txt

### Run Streamlit app
streamlit run app.py

## IMPORT LIBRARIES

import os
import numpy as np
import pandas as pd

#Automcomplete
%config IPCompleter.greedy=True

## DIRECTORY

root = '/Users/rober/'

### Project name

dir_name = 'smartport-delay-risk-scoring'

### Create the directory and project structure

path = root + dir_name

try:
    os.mkdir(path)
    os.mkdir(path + '/01_Documents')
    os.mkdir(path + '/02_Data')
    os.mkdir(path + '/02_Data/01_Raw')
    os.mkdir(path + '/02_Data/02_Validation')
    os.mkdir(path + '/02_Data/03_Working')
    os.mkdir(path + '/02_Data/04_Caches')
    os.mkdir(path + '/03_Notebooks')
    os.mkdir(path + '/03_Notebooks/01_Functions')
    os.mkdir(path + '/03_Notebooks/02_Development')
    os.mkdir(path + '/03_Notebooks/03_System')
    os.mkdir(path + '/04_Models')
    os.mkdir(path + '/05_Outputs')
    os.mkdir(path + '/09_Other')
    
except OSError:
    print ("The directory %s has NOT been created" % path)
else:
    print ("The directory %s has been succesfully created" % path)

## IMPORT DATA

file_name_data = 'tracking_db.csv' 

path_data_raw = path + '/02_Data/01_Raw/' + file_name_data 

df = pd.read_csv(path_data_raw, low_memory=False)
df

The dataset contains AIS-like (Automatic Identification System) tracking records for container ships travelling between ports.

**Each row** represents a single **position update** for a ship, including timing, location, movement data and scheduled / actual departure and arrival information.



| **Field**       | **Description**                                                             |
| --------------- | --------------------------------------------------------------------------- |
| **id**          | Unique identifier for each data record                                      |
| **updated**     | Timestamp indicating when the record was last updated                       |
| **ship**        | Vessel name                                                                 |
| **imo**         | International Maritime Organization (IMO) number — unique vessel identifier |
| **lat**         | Latitude coordinate of vessel position                                      |
| **long**        | Longitude coordinate of vessel position                                     |
| **sog**         | Speed Over Ground (knots) — vessel speed relative to Earth                  |
| **cog**         | Course Over Ground (degrees) — actual direction of vessel movement          |
| **hdg**         | Heading (degrees) — direction the vessel’s bow is pointing                  |
| **depPort**     | Departure port code                                                         |
| **etdSchedule** | Scheduled Estimated Time of Departure                                       |
| **etd**         | Updated Estimated Time of Departure (may differ from schedule)              |
| **atd**         | Actual Time of Departure                                                    |
| **arrPort**     | Arrival port code                                                           |
| **etaSchedule** | Scheduled Estimated Time of Arrival                                         |
| **eta**         | Updated Estimated Time of Arrival (may differ from schedule)                |
| **ata**         | Actual Time of Arrival                                                      |




These fields will be used to build voyage-level features and to derive the target variable indicating whether a vessel experienced a relevant arrival delay.


### Rename columns

df = df.rename(columns={
    "id":"record_id",
    "updated":"updated_ts",
    "ship":"ship_name",
    "imo":"imo",
    "lat":"lat",
    "long":"lon",
    "sog":"sog",
    "cog":"cog",
    "hdg":"hdg",
    "depPort":"dep_port",
    "etdSchedule":"etd_schedule",
    "etd":"etd",
    "atd":"atd",
    "arrPort":"arr_port",
    "etaSchedule":"eta_schedule",
    "eta":"eta",
    "ata":"ata",
})

df

### Create validation dataset

val = df.sample(frac=0.3, random_state=42)

validation_file = 'validation.csv'
validation_path = path + '/02_Data/02_Validation/' + validation_file

val.to_csv(validation_path, index=False)

### Create working dataset

work = df.loc[~df.index.isin(val.index)]

work_file = 'work.csv'
work_path = path + '/02_Data/03_Working/' + work_file

work.to_csv(work_path, index=False)

### Create a sample

sample = work.sample(n=20000, random_state=42)

sample_file = 'sample.csv'
sample_path = path + '/02_Data/03_Working/' + sample_file

sample.to_csv(sample_path, index=False)
